+++ src/main/java/com/wenxin2/marioverse/datagen/ItemModelGen.java
@@ -98,6 +98,7 @@ public class ItemModelGen extends ItemModelProvider {
         this.basicItem(ItemRegistry.PLUMBER_POTTERY_SHERD.get());
         this.basicItem(ItemRegistry.POKEY_SPAWN_EGG.get());
         this.basicItem(ItemRegistry.PORCUPUFFER.get());
+        this.basicItem(ItemRegistry.PRINCESS_ARMOR_TRIM_SMITHING_TEMPLATE.get());
         this.basicItem(ItemRegistry.QUICKSAND_BUCKET.get());
         this.basicItem(ItemRegistry.RED_KOOPA_SHELL.get());
         this.basicItem(ItemRegistry.RED_KOOPA_SHOES.get());