+++ src/main/java/com/wenxin2/marioverse/datagen/ItemTagsGen.java
@@ -260,7 +260,8 @@ public class ItemTagsGen extends ItemTagsProvider {
 
         tag(ItemTags.TRIM_TEMPLATES)
                 .add(ItemRegistry.LUIGI_ARMOR_TRIM_SMITHING_TEMPLATE.get())
-                .add(ItemRegistry.MARIO_ARMOR_TRIM_SMITHING_TEMPLATE.get());
+                .add(ItemRegistry.MARIO_ARMOR_TRIM_SMITHING_TEMPLATE.get())
+                .add(ItemRegistry.PRINCESS_ARMOR_TRIM_SMITHING_TEMPLATE.get());
 
         tag(ItemTags.TRIMMABLE_ARMOR)
                 .add(ItemRegistry.BODICE.get())